Separate SlimMedia components into separate model file
